+++ CHANGELOG.md
@@ -203,6 +203,8 @@ Encounter typescipt bug, VSC 1.94.2, Node.js 20.16.0, Npm 10.8.2
 
 ## 0.12.2024120502
 
+- Fix view for staged changes
+
 - Icon for project tasks title
 
 - Icon for project tasks