+++ CHANGELOG.md
@@ -205,6 +205,4 @@ Encounter typescipt bug, VSC 1.94.2, Node.js 20.16.0, Npm 10.8.2
 
 - Fix view for staged changes
 
-- Icon for project tasks title
-
-- Icon for project tasks
+- Git button shown depend on git state