+++ CHANGELOG.md
@@ -206,3 +206,7 @@ Encounter typescipt bug, VSC 1.94.2, Node.js 20.16.0, Npm 10.8.2
 - Fix view for staged changes
 
 - Git button shown depend on git state
+
+## 0.12.2024120601
+
+- Fix icons for smart tasks tree view